Update SpanTask. Update SpanCatTask.
