+++ spacy_llm/tasks/lemma/parser.py
@@ -1,7 +1,7 @@
 from typing import Iterable, List
 
 
-def parse_responses_v1(responses: Iterable[str]) -> Iterable[List[List[str]]]:
+def parse_responses_v1(responses: Iterable[str], **kwargs) -> Iterable[List[List[str]]]:
     """Parses LLM responses for spacy.Lemma.v1.
     responses (Iterable[str]): LLM responses.
     RETURNS (Iterable[List[str]]): Lists of 2-lists (token: lemmatized token) per doc/response.
