+++ spacy_llm/tasks/sentiment/task.py
@@ -5,7 +5,7 @@ from spacy.language import Language
 from spacy.tokens import Doc
 from spacy.training import Example
 
-from ...ty import FewshotExample, TaskResponseParserType
+from ...ty import FewshotExample, TaskResponseParserProtocol
 from ..templates import read_template
 from ..util import SerializableTask
 from .examples import SentimentExample
@@ -17,7 +17,7 @@ class SentimentTask(SerializableTask):
     def __init__(
         self,
         template: str,
-        parse_responses: TaskResponseParserType,
+        parse_responses: TaskResponseParserProtocol,
         fewshot_example_type: Type[FewshotExample],
         field: str,
         examples: Optional[List[SentimentExample]],
