+++ spacy_llm/tasks/span/__init__.py
@@ -1,0 +1,5 @@
+from .examples import SpanExample
+from .parser import parse_responses_v1_v2
+from .task import SpanTask
+
+__all__ = ["parse_responses_v1_v2", "SpanExample", "SpanTask"]
