+++ spacy_llm/tasks/span/examples.py
@@ -1,0 +1,9 @@
+import abc
+from typing import Dict, List
+
+from ...ty import FewshotExample
+
+
+class SpanExample(FewshotExample, abc.ABC):
+    text: str
+    entities: Dict[str, List[str]]
